use ffmpeg-sys-next, single version of bindgen (#162)
* use ffmpeg-sys-next, single version of bindgen

One version is easier to keep track of and should reduce build times.

* replace ffmpeg-sys with ffmpeg-sys-next, which (among other things)
  uses bindgen 0.64. ffmpeg-sys was using a pre-0.60 version, so in
  some of our call sites, it hit build errors on Rust 1.70 due to
  https://github.com/rust-lang/rust-bindgen/issues/2083
  https://github.com/rust-lang/rust-bindgen/blob/main/CHANGELOG.md#0600

* update cargo lock stuff to use the same version

* remove stale comments about older versions

* fix xilinx build errors with bindgen 0.64

+++ xilinx/src/xlnx_enc_props.rs
@@ -355,7 +355,7 @@ impl XlnxXmaEncoderProperties {
     ) {
         enc_params[0].name = ENC_OPTIONS_PARAM_NAME.as_ptr() as *mut i8;
         enc_params[0].type_ = XmaDataType_XMA_STRING;
-        enc_params[0].length = enc_options.len() as u64;
+        enc_params[0].length = enc_options.len();
         enc_params[0].value = enc_options_ptr_ptr as *mut std::ffi::c_void;
 
         enc_params[1].name = LATENCY_LOGGING_PARAM_NAME.as_ptr() as *mut i8;
